requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
jinja2
groq
langchain
langchain-community
openai
sse-starlette
python-dotenv